Create README.md
Crapstruder - When life gives you Afinia parts, build a crapstruder!

I found myself with a spare Afinia/UP! extruder, and I wanted to be able use the motor and supporting hardware with a different hotend on a different printer. 

This should let you use that motor+extruder hardware with a Jhead-compatible hotend on a Prusa i3 "wade" style x-carriage.